refactor: add modular app skeleton
