+++ gpt2giga/api/__init__.py
@@ -1,0 +1,1 @@
+"""Future public API router namespace."""
