+++ gpt2giga/api/anthropic/__init__.py
@@ -1,0 +1,5 @@
+"""Anthropic-compatible API router namespace."""
+
+from gpt2giga.api.anthropic.routes import router
+
+__all__ = ["router"]
