+++ gpt2giga/api/openai/__init__.py
@@ -1,0 +1,5 @@
+"""OpenAI-compatible API router namespace."""
+
+from gpt2giga.api.openai.routes import router
+
+__all__ = ["router"]
