+++ gpt2giga/api/system/__init__.py
@@ -1,0 +1,1 @@
+"""System API router namespace."""
