+++ gpt2giga/app/__init__.py
@@ -1,0 +1,1 @@
+"""Application factory and lifecycle namespace."""
