+++ gpt2giga/protocols/__init__.py
@@ -1,0 +1,1 @@
+"""Protocol adapter namespace for the modular gateway path."""
