+++ gpt2giga/protocols/anthropic/__init__.py
@@ -1,0 +1,1 @@
+"""Anthropic protocol adapter namespace."""
