+++ gpt2giga/protocols/normalized/__init__.py
@@ -1,0 +1,1 @@
+"""Normalized protocol model namespace."""
