+++ gpt2giga/protocols/openai/__init__.py
@@ -1,0 +1,1 @@
+"""OpenAI protocol adapter namespace."""
