+++ gpt2giga/providers/gigachat/__init__.py
@@ -1,0 +1,17 @@
+"""GigaChat provider adapter namespace."""
+
+from gpt2giga.providers.gigachat.auth import (
+    create_gigachat_client_for_request,
+    pass_token_to_gigachat,
+)
+from gpt2giga.providers.gigachat.client import (
+    close_gigachat_client,
+    create_gigachat_client,
+)
+
+__all__ = [
+    "close_gigachat_client",
+    "create_gigachat_client",
+    "create_gigachat_client_for_request",
+    "pass_token_to_gigachat",
+]
