+++ gpt2giga/sinks/__init__.py
@@ -1,0 +1,1 @@
+"""Extension sink namespace."""
