+++ gpt2giga/sinks/logs/__init__.py
@@ -1,0 +1,1 @@
+"""Traffic log sink namespace."""
