+++ gpt2giga/sinks/metrics/__init__.py
@@ -1,0 +1,1 @@
+"""Metrics sink namespace."""
