+++ gpt2giga/sinks/observability/__init__.py
@@ -1,0 +1,1 @@
+"""Observability sink namespace."""
