+++ tests/test_api_server/test_ci_smoke.py
@@ -74,7 +74,10 @@ def make_app(monkeypatch):
             gigachat_api_mode="v1",
         )
     )
-    monkeypatch.setattr("gpt2giga.api_server.GigaChat", lambda **kw: FakeGigaChat())
+    monkeypatch.setattr(
+        "gpt2giga.app.lifecycle.create_gigachat_client",
+        lambda settings: FakeGigaChat(),
+    )
     return create_app(config=config)
